Change macro to check for an existing tupled
